v.3.0.0-beta - удаление jquery.coockie.js - utm-метки на js получаются без jquery - плагин переделан на новую структуру Joomla 4 - новая таблица связей в базу данных для товаров JoomShopping и Битрикс 24. - возможность сопоставить товары JoomShopping и Битрикс 24 в карточке товара - Складской учёт: выбор склада по умолчанию - передача товаров из заказа JoomShopping  в Битрикс 24 не только товарными позициями, но и товарами (участвуют в аналитике Битрикс 24, резервируются со склада и т.д.) - обновление цен и остатков товаров JoomShopping при редактировании в Битрикс 24.
