--- img/index.html
@@ -1,1 +1,0 @@
-<html><body bgcolor="#FFFFFF"></body></html>
