--- language/en-GB/index.html
@@ -1,4 +1,0 @@
-<html>
-<head></head>
-<body></body>
-</html>
