+++ language/ru-RU/ru-RU.plg_system_wt_jshopping_b24_pro.sys.ini
@@ -4,4 +4,4 @@ PLG_WT_JSHOPPING_B24_PRO_AFTER_INSTALL = "Спасибо за установку
 PLG_WT_JSHOPPING_B24_PRO_AFTER_DISCOVER_INSTALL = "Спасибо за установку расширения"
 PLG_WT_JSHOPPING_B24_PRO_AFTER_UPDATE = "Спасибо за обновление расширения"
 PLG_WT_JSHOPPING_B24_PRO_MAYBE_INTERESTING="<h4>Может быть интересно</h4><ul class=\"nav nav-tabs nav-stacked\"><li><a href=\"https://web-tolk.ru/dev/joomla-modules/wt-quick-links.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\"><span class=\"label label-success\">New</span> <strong>WT Quick links</strong> - Модуль для быстрого вывода ссылок на категории материалов, Virtuemart, JoomShopping, пункт меню или пользовательскую ссылку. Разрабатывался для карусели ссылок-тегов в категориях товаров.</a></li><li><a href=\"https://web-tolk.ru/dev/joomla-plugins/wt-seo-meta-templates.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\"><span class=\"label label-success\">New</span> <strong>WT SEO Meta templates</strong> - бесплатный плагин сео-шаблонов для &lt;title&gt; и meta-description<br/>в материалах и категория Joomla, Virtuemart, My City Selector.</a></li><li><a href=\"https://web-tolk.ru/dev/joomshopping/wt-sm-otpravka-pochta-ru.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\"><span class=\"label label-info\">Paid</span> <strong>WT SM Otpravka.pochta.ru</strong> - интеграция JoomShopping с Почтой России для Бизнеса</a></li><li><a href=\"https://web-tolk.ru/dev/joomshopping/wt-joomshopping-favorite.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\"><strong>WT JoomShopping Favorites</strong> - альтернативный список пожеланий (wishlist) для JoomShopping.</a></li></ul>"
-PLG_WT_JSHOPPING_B24_PRO_WHATS_NEW="<h3>v.2.6.0 Что нового?</h3><h4>Косметические изменения</h4><p><ul><li>Убраны лишние разделители в комментарии к лиду/сделке</li><li>Появилась возможность отключить показ ссылки на просмотр заказа в JoomShopping в комментарии лида/сделки</li><li>Название ссылки получило языковую константу и отображается на русском языке</li></ul></p><h4>Выбор условия создания лида/сделки</h4><p>Теперь можно выбрать при каком условии создавать лид/сделку: <ul><li><strong>Всегда</strong> (режим по умолчанию) - в Битрикс24 попадают как оплаченные, так и не завершённые заказы, однако у менеджеров по продажам есть возможность узнать проблемы клиента и \"дожать\" продажу.</li><li><strong>После успешной оплаты</strong> - обратите внимание, что событие успешной оплаты учитывается только в том случае, если покупатель вернулся в магазин из платёжной системы. Некоторые покупатели закрывают страницу платёжной системы сразу после оплаты, не возвращаясь в магазин. Также в режиме только успешной оплаты не будет поддерживаться создание лидов/сделок с помощью плагина Nevigen Quick Order.</li></ul></p><h4>Добавлено новое значение для отправки в Битрикс24 - трек-номер Почты России</h4><p>Трек-номер создается с помощью расширения WT SM Otpravka.pochta.ru. В группе системных плагинов важно установить порядок срабатывания плагинов так, чтобы плагин интеграции с Почтой России срабатывал раньше, чем отправка заказа в Битрикс24.</p><p><h4>Демо видео настройки двухсторонней интеграции</h4><iframe width="560" height="315" src=\"https://www.youtube.com/embed/6Uo3LEnKJ2g\" title=\"YouTube video player\" frameborder=\"0\" allow=\"accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture\" allowfullscreen></iframe></p>"
+PLG_WT_JSHOPPING_B24_PRO_WHATS_NEW="<h3 class="text-danger">v.3.0.0-alpha Что нового?</h3><h4>Косметические изменения</h4><p><ul><li>Убраны лишние разделители в комментарии к лиду/сделке</li><li>Появилась возможность отключить показ ссылки на просмотр заказа в JoomShopping в комментарии лида/сделки</li><li>Название ссылки получило языковую константу и отображается на русском языке</li></ul></p><h4>Выбор условия создания лида/сделки</h4><p>Теперь можно выбрать при каком условии создавать лид/сделку: <ul><li><strong>Всегда</strong> (режим по умолчанию) - в Битрикс24 попадают как оплаченные, так и не завершённые заказы, однако у менеджеров по продажам есть возможность узнать проблемы клиента и \"дожать\" продажу.</li><li><strong>После успешной оплаты</strong> - обратите внимание, что событие успешной оплаты учитывается только в том случае, если покупатель вернулся в магазин из платёжной системы. Некоторые покупатели закрывают страницу платёжной системы сразу после оплаты, не возвращаясь в магазин. Также в режиме только успешной оплаты не будет поддерживаться создание лидов/сделок с помощью плагина Nevigen Quick Order.</li></ul></p><h4>Добавлено новое значение для отправки в Битрикс24 - трек-номер Почты России</h4><p>Трек-номер создается с помощью расширения WT SM Otpravka.pochta.ru. В группе системных плагинов важно установить порядок срабатывания плагинов так, чтобы плагин интеграции с Почтой России срабатывал раньше, чем отправка заказа в Битрикс24.</p><p><h4>Демо видео настройки двухсторонней интеграции</h4><iframe width="560" height="315" src=\"https://www.youtube.com/embed/6Uo3LEnKJ2g\" title=\"YouTube video player\" frameborder=\"0\" allow=\"accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture\" allowfullscreen></iframe></p>"
