+++ script.php
@@ -32,11 +32,11 @@ class plgSystemWt_jshopping_b24_proInstallerScript
     /**
      * This method is called after a component is installed.
      *
-     * @param  \stdClass $parent - Parent object calling this method.
+     * @param  \stdClass $installer - Parent object calling this method.
      *
      * @return void
      */
-    public function install($parent)
+    public function install($installer)
     {
 
     }
@@ -44,11 +44,11 @@ class plgSystemWt_jshopping_b24_proInstallerScript
     /**
      * This method is called after a component is uninstalled.
      *
-     * @param  \stdClass $parent - Parent object calling this method.
+     * @param  \stdClass $installer - Parent object calling this method.
      *
      * @return void
      */
-    public function uninstall($parent) 
+    public function uninstall($installer) 
     {
 
 		
@@ -57,11 +57,11 @@ class plgSystemWt_jshopping_b24_proInstallerScript
     /**
      * This method is called after a component is updated.
      *
-     * @param  \stdClass $parent - Parent object calling object.
+     * @param  \stdClass $installer - Parent object calling object.
      *
      * @return void
      */
-    public function update($parent) 
+    public function update($installer) 
     {
 
     }
@@ -74,11 +74,11 @@ class plgSystemWt_jshopping_b24_proInstallerScript
      *                           - * install
      *                           - * update
      *                           - * discover_install
-     * @param  \stdClass $parent - Parent object calling object.
+     * @param  \stdClass $installer - Parent object calling object.
      *
      * @return void
      */
-    public function preflight($type, $parent) 
+    public function preflight($type, $installer) 
     {
 
     }
@@ -92,13 +92,21 @@ class plgSystemWt_jshopping_b24_proInstallerScript
      *                           - * install
      *                           - * update
      *                           - * discover_install
-     * @param  \stdClass $parent - Parent object calling object.
+     * @param  \stdClass $installer - Parent object calling object.
      *
      * @return void
      */
-    function postflight($type, $parent) 
+    function postflight($type, $installer)
     {
+	    $smile = '';
+	    if($type != 'uninstall')
+	    {
+		    $smiles    = ['&#9786;', '&#128512;', '&#128521;', '&#128525;', '&#128526;', '&#128522;', '&#128591;'];
+		    $smile_key = array_rand($smiles, 1);
+		    $smile     = $smiles[$smile_key];
+	    }
 
+	    $element = strtoupper($installer->getElement());
 		echo "
 		<style>	.thirdpartyintegration {
 				display:flex;
@@ -120,13 +128,13 @@ class plgSystemWt_jshopping_b24_proInstallerScript
 				background-color:#ffdddb;
 			}
 		</style>
-		<div class='row' style='margin:25px auto; border:1px solid rgba(0,0,0,0.125); box-shadow:0px 0px 10px rgba(0,0,0,0.125); padding: 10px 20px;'>
-		<div class='span8 control-group' id='wt_download_id_form_wrapper'>
-		<h2>".JText::_("PLG_".strtoupper($parent->get("element"))."_AFTER_".strtoupper($type))." <br/>".JText::_("PLG_".strtoupper($parent->get("element")))."</h2>
-		".Text::_("PLG_".strtoupper($parent->get("element"))."_DESC");
+		<div class='row bg-white m-3 p-3 shadow-sm border'>
+		<div class='col-12 col-lg-8'>
+		<h2>".$smile." ".Text::_("PLG_".$element."_AFTER_".strtoupper($type))." <br/>".Text::_("PLG_".$element)."</h2>
+		".Text::_("PLG_".$element."_DESC");
 		
 		
-			echo JText::_("PLG_".strtoupper($parent->get("element"))."_WHATS_NEW");
+			echo Text::_("PLG_".$element."_WHATS_NEW");
 
 
 		    $thirdpartyextensions="";
@@ -166,11 +174,16 @@ class plgSystemWt_jshopping_b24_proInstallerScript
 		    }
 
 		echo "</div>
-		<div class='span4' style='display:flex; flex-direction:column; justify-content:center;'>
+		<div class='col-12 col-lg-4 d-flex flex-column justify-content-start'>
 		<img width='200px' src='https://web-tolk.ru/web_tolk_logo_wide.png'>
 		<p>Joomla Extensions</p>
-		<p><a class='btn' href='https://web-tolk.ru' target='_blank'><i class='icon-share-alt'></i> https://web-tolk.ru</a> <a class='btn' href='mailto:info@web-tolk.ru'><i class='icon-envelope'></i>  info@web-tolk.ru</a></p>
-		".JText::_("PLG_".strtoupper($parent->get("element"))."_MAYBE_INTERESTING")."
+		<p class='btn-group'>
+			<a class='btn btn-sm btn-outline-primary' href='https://web-tolk.ru' target='_blank'>https://web-tolk.ru</a>
+			<a class='btn btn-sm btn-outline-primary' href='mailto:info@web-tolk.ru'><i class='icon-envelope'></i> info@web-tolk.ru</a>
+		</p>
+		<p><a class='btn btn-info' href='https://t.me/joomlaru' target='_blank'>Joomla Russian Community in Telegram</a></p>
+		
+		".Text::_("PLG_".$element."_MAYBE_INTERESTING")."
 		</div>
