+++ sql/install.mysql.utf8.sql
@@ -5,3 +5,8 @@ CREATE TABLE IF NOT EXISTS `#__wt_jshopping_bitrix24_pro` (
   `bitrix24_deal_id` int(11),
   PRIMARY KEY (`id`)
 ) DEFAULT CHARSET=utf8 ;
+
+CREATE TABLE IF NOT EXISTS `#__wt_jshopping_bitrix24_pro_products_relationship` (
+  `jshopping_product_id` int(11) UNIQUE,
+  `bitrix24_product_id` int(11)
+) DEFAULT CHARSET=utf8 ;
