+++ sql/uninstall.mysql.utf8.sql
@@ -1,1 +1,2 @@
 DROP TABLE #__wt_jshopping_bitrix24_pro ;
+DROP TABLE #__wt_jshopping_bitrix24_pro_products_relationship ;
